+++ src/utils/common/constants.ts
@@ -167,7 +167,7 @@ export const constants = {
   secretsScanner: "Checkmarx Secret Detection Realtime Scanner",
   secretsScannerStart: "Secret Detection Scanner Engine started",
   secretsScannerDisabled: "Secret Detection Scanner Engine disabled",
-  secretsScannerEngineName: "SecretScanner",
+  secretsScannerEngineName: "Secrets",
   secretsScannerDirectory: "Cx-secret-realtime-scanner",
   errorSecretsScanRealtime: "Failed to handle Secret Detection scan",
 
@@ -176,7 +176,7 @@ export const constants = {
   ossRealtimeScanner: "Checkmarx Open Source Realtime Scanner (OSS-Realtime)",
   ossRealtimeScannerStart: "Realtime OSS Scanner Engine started",
   ossRealtimeScannerDisabled: "Realtime OSS Scanner Engine disabled",
-  ossRealtimeScannerEngineName: "RealtimeOssScanner",
+  ossRealtimeScannerEngineName: "Oss",
   ossRealtimeScannerDirectory: "Cx-oss-realtime-scanner",
   errorOssScanRealtime: "Failed to handle OSS Realtime scan",
 
@@ -245,6 +245,9 @@ export const constants = {
   copilotChatExtensionId: "GitHub.copilot-chat",
   copilotNewChatOpen: "workbench.action.chat.newChat",
   copilotChatOpenWithQueryCommand: "workbench.action.chat.openAgent",
+
+  openAIChat: "fixWithAIChat",
+  viewDetails: "viewPackageDetails"
 };
 
 export enum Platform {